Cover multi-dof joints and unify waypoint control
The initial waypoint rewrite only bounded REVOLUTE and PRISMATIC
joints, silently freezing every D6 joint the MJCF importer creates by
merging a body's hinges - on the humanoid 16 of 21 actuated dofs.
Bounds now cover every joint whose coordinates map 1:1 to dofs and
skip only quaternion-bearing joints; sentinel or inverted limit ranges
fall back to the initial target instead of adopting huge or reversed
bounds.

The host and device paths also duplicated the slew logic with
different RNGs, and the host path drew leg-0 waypoints from the same
PCG64(123) stream that create_model_builder uses for initial-state
randomization. Both paths now launch one shared warp kernel
(graph-captured by the Fast benchmarks, enqueued per frame by the Kpi
path outside the timed region), which drops the per-frame host upload
and the now-unused Example.rng.
